minimint-api/ln: interconnect modules using generalized API
The generalized API allows modules to exchange consensus state.
This is useful for the LN module which needs access to consensus height but should not have to run duplicate consensus on that.
By standardizing the module API modules can now query each other's APIs without opening fallible network connections which would be an unacceptable risk.