Catch panics in API handlers
Previously, when the API was using tide as its server, panics were
caught and returned as HTTP errors. This means that this might still be
an implicit assumption in some places and should be investigated.
Catching panics is only the last line of defense, but it seems prudent.

Related to #398