Rename `minimint` the binary to `fedimintd`
This is a bit of a subset of #293. I think it is going to be
easier to merge such change in chunks than a one big whole
that generates merge conficts with everything.

And we care about this particular binary name because we're
setting up OCI container builds in Dockre Hub and it's
easier to start with the desired name upfront than change
it later.