Introduce cargo feature `bitcoind` for integration tests
This cargo feature adds the ability for integration tests to spawn a
child process of `bitcoind` (via the `bitcoind` crate).

When the `bitcoind` cargo feature is enabled, internally-spawned
`bitcoind` will replace mock rpc client and rpc client that interacts
with externally-spawned `bitcoind`.

A new env `BITCOIN_EXE` is added which the `bitcoind` binary location
can be specified. If unspecified, the `bitcoind` crate will download
the binary.

`RealBitcoinTest` is now an enum, either specifying just a client, or
both a service and a client (with `bitcoind::BitcoinD`).