Introduce "dyn newtype" pattern and convert `trait Database` to it
Since we are moving more and more things to use dynamic dispatch,
it might be worthwhile to introduce this pattern early, as
I am already planning to use use it extensively in modularization
work.

I've selected `Database` as a first trait to convert to give this
approach a good test.