Use nightly rustfmt
I was reading comments under recent announcement of creation of
Rust Style Team:

https://www.reddit.com/r/rust/comments/xreg1a/announcing_the_rust_style_team/

and someone mentioned:

```
group_imports = StdExternalCrate
```

and it seemed like such a nice way to format imports, but available only
on nightly rust.

So I wondered if it would be possible to use stable compiler, but nightly
`rustmft` and it seems it would work.

Presenting for consideration.