Change `Decodable<M>` to `Decodable`
After a while of working with the current version of Decodable
I've hit a wall, with not being able to express:

```
trait Smth : for <M : Decoder> Decodable<M>
```

without some heavy HKT machiner.

This led to a realization that the current form is conceptually
wrong. I don't need a trait that is parametrized by `M`.
I need one trait, that when impleemnted guarantee that any `M`
(with a bound) can be used.