Move `core` into `fedimint-api`
for `Decodable` macro to be usable, I need to put a bound on
`Decodable::consensus_decode`: `M : ModuleDecoder`, and such
`ModuleDecoder` by necessity will need to refer to types it
can decode.

So I don't see a way to avoid just bringing this `core`
code under `fedimint-api`. That's also how the end goal
is going to look like (more or less) anyway, so might as well
do it already.