Make module errors abstract
Currently `FedimintModule` trait uses errors that are imported from
the specific modules. This is not possible in a post-modularization
code.

As a initial step toward modularization, make all the errors returned
by the server side module abstract (anyhow::Error).

In the future we might want to specify some concrete Error variants
with `Other` being just an escape hatch.