Add "admin key" for guardians to sign rpcs to fedimintd
We are going to need a way for the guardians to do things remotely
via their app.

For social recovery the guardian needs to inform the fedimintd to
create and publish decryption share, ideally from the conveniance
of their app.

In the future we might want an option for the fedimintd to start in
a locked mode and require remote config decryption (so we don't have
to store any passphrase locally at all).

Other modules might require their own guardian authentication.