Get rid of bitcoind_rpc as configuration of `fedimint-wallet`
The wallet module already gets the API (`dyn Trait`) injected.
It does not need to know by itself where is the bitcoin data
comming from. It simplifies a lot of other things, where
we don't need to fake things we don't need anyway.

The end application is responsible for providing wallet module
with blockchain data: be that bitcoind rpc client, or
(in the future electrum client).

From an operational PoV, the bitcoind rpc port is also so
important, that it's best to bring it upfront instead of
hidding in a module config.