Split `ModuleKey` into `ModuleInstanceId` & `ModuleKind`
So far we've been conflating the two, thinking that a federation
will run each module type exactly once, using a `u16` value
as a "globally unique ID".

After the modularization is complete, it became more clear
that actually we want a seprate id for a "module type"
(named here "kind" to avoid confussion with Rust "type"),
and a separate id for each module federation decided to
"instantiate".

Each fedimint binary (like `fedimintd` will register all available
module kinds) early in `main` function.

```rust
fn main() {
    FedimintServer::builder()
        .datadir("/var/lib/fedimintd")
        .with_module::<WalletInit>()
        .with_module::<MintInit>()
        .with_module::<LightningInit>()
        .build()
        .run()
        .await
}
```

The the users will define module instances they actually want to run
as a part of the federation (we don't need this part yet, and we can
just assume each module kind is insantiated exactly once).

The DKG will generate federation config that will have module configs
keyed by the `ModuleInstanceId`, and containing `kind: ModuleKind`
for each instance, so like:

```json
{
  // ...

  modules: {
    1: {
      kind: "mint",
      // ..
    },
    2: {
      kind: "wallet",
    }
  }
  // ...
}
```

On start, when config is initialized `fedimintd` will use the config
to know for which id, to instantiate what module with what config.

In the future we might want to implement ability to change the
composition of module instances. So the Federation can disable certain
module, and replace it with a different one, etc.

It's undecided yet, are we going to use it for module version change
coordination, or just to give ability to run same module multiple
times with different configs (might be useful for 3rd party modules).

One way or another the architecture seems better if the id of the
instance is separate from it's "type".

This change only does the minimum rename and is to gather approval
for the idea. In follow-up changes we will have to actually
replace certain parts of code-base that are using strings, where
they should be using ids, etc.

Split `ModuleKey` into `ModuleInstanceId` & `ModuleKind`

So far we've been conflating the two, thinking that a federation
will run each module type exactly once, using a `u16` value
as a "globally unique ID".

After the modularization is complete, it became more clear
that actually we want a seprate id for a "module type"
(named here "kind" to avoid confussion with Rust "type"),
and a separate id for each module federation decided to
"instantiate".

Each fedimint binary (like `fedimintd` will register all available
module kinds) early in `main` function.

```rust
fn main() {
    FedimintServer::builder()
        .datadir("/var/lib/fedimintd")
        .with_module::<WalletInit>()
        .with_module::<MintInit>()
        .with_module::<LightningInit>()
        .build()
        .run()
        .await
}
```

The the users will define module instances they actually want to run
as a part of the federation (we don't need this part yet, and we can
just assume each module kind is insantiated exactly once).

The DKG will generate federation config that will have module configs
keyed by the `ModuleInstanceId`, and containing `kind: ModuleKind`
for each instance, so like:

```json
{
  // ...

  modules: {
    1: {
      kind: "mint",
      // ..
    },
    2: {
      kind: "wallet",
    }
  }
  // ...
}
```

On start, when config is initialized `fedimintd` will use the config
to know for which id, to instantiate what module with what config.

In the future we might want to implement ability to change the
composition of module instances. So the Federation can disable certain
module, and replace it with a different one, etc.

It's undecided yet, are we going to use it for module version change
coordination, or just to give ability to run same module multiple
times with different configs (might be useful for 3rd party modules).

One way or another the architecture seems better if the id of the
instance is separate from it's "type".

This change only does the minimum rename and is to gather approval
for the idea. In follow-up changes we will have to actually
replace certain parts of code-base that are using strings, where
they should be using ids, etc.