feat: Fetch epochs during ecash recovery on 8 threads
It seems that in practice fetching epochs take a long time leading
to a bad user-experience. Fetching epochs from federation has a
large latency, so it makes sense to do in on multiple-threads.

I selected 8 as a good compromise that should not create issues,
like too much loads, while hopefully giving most of the benefit
of paralellism.