feat: spell-checking via `cargo spellcheck`
I got curious and wanted to try it out, but I'm not sure if it's
worth following through.

First, `cargo spellcheck` seem rather buggy and slow.

Second, it's a lot of effort. Hunspell doesn't know *a lot* of
the terms we're using.

I rather get some feedback before investing any more time in it.