fix: upload & restore ecash by requesting it from more peers
`ecash_can_be_recovered` is flaky. The reason is that sometimes the
test can't find a snapshot it made at the very begining, and due
to the hacky nature of how test user is funded, without the snapshot
with initial funds we can't really do the recovery.

The reason we can't find snapshot is that test federation is made
of 2 peers only. Requesting backup with `one_honest` makes it
to one federation member while requesting recovery snapshot
can finish after the other peer responds with "None".

It seems that for backup purposes it's better to require more
responses.