refactor: split `impl_db_record` into two macros
Why? Because the types these macro `impl`s traits for
can come from different crates making it impossible to use it.

Currently this does not happen, but I'm experimenting with 3rd party
module building and spliting module into server/client/common sub-parts
and this becomes a blocker.

Also, it seems it's easier to use and understand when each macro is
smaller (functionaly).