refactor: embedded the consensus hash in client config module entry
The goal is to support clients that don't have the support for
all the modules that federation is running, which will often be the
case for LN gateways.

To facilitate that we embed the hash the config itself, so that
the module that doesn't support it, can just use that value. As
they don't support it, they don't really care if if the config
value for that module is valid - they are not going to use it
anyway.

Since this is a second extra key in client module config object
turn:

```
{
  kind: "kind",
  ... rest of the config
}
```

into:

```
 {
  kind: "kind",
  hash: "hash",
  config: { ... rest of the config }
}
```

which eliminates a potential duplicate key problem.