fix: shut down CLN plugin when lightningd stops running
There are 2 cases where we shutdown and cleanup our CLN plugin:
1. When user requests a shutdown of the plugin (e.g. via `lightning-cli
plugin stop`
2. When lightningd stops running for whatever reason

In the case of (1), our plugin receives a "shutdown" notification. We
handle this by calling `plugin.shutdown()` which causes the
`plugin.join()` future to resolve.

In the case of (2), `cln_plugin` will read a 0-byte message from STDIN
and trigger the `plugin.join()` future to resolve.

With this change we can handle all cleanup logic in one place: when
`plugin.join()` resolve.