refactor: ServerModule::distributed_gen interface (#1749)
We're passing a lot of arguments there, all which should not be
accesible to the module itself. The module should have no access
to its instance id, or raw dkg connection.

I think we'll need to define more operations, and possibly
refactor it all even more, but this should be a good starting
step in the right direction.