fix: can't use `fedimint-cli decode-connect-info` without `--workdir`
Refactor `fedimint-cli` to get rid of the `--workdir` hacks.

The root of the problem is that we don't know if we need a `Client`
(and thus `--workdir`) before we know which sub-command we are handling.
So lets accept that fact, and just build the client dynamically when
we need it, instead of trying to maintain a list of places where we
actually don't need it, etc.