chore: make Code Coverage build run tests in parallel
According to:

https://github.com/taiki-e/cargo-llvm-cov#known-limitations

it's only defaulting to 1 thread, because of rustc issue:

https://github.com/rust-lang/rust/issues/91092

but it seems the issue is that relatively infrequently some tests
will fail to be reported... which if fine with me if it makes
the CI faster. And they are talking about thousands of tests,
while we probably have <100.