chore: don't run *all* cross-compilation targets on macos
We only had a handful of macos runners available, and they take
a while to build their stuff. Also - due to how filtering here
needs to be done, even builds that we skip need to get an available
macos runner, just so it can... ignore the step.

All this combined, stals PR's CI, due to not enough macos runners.