feat: update rust code so old and new setup UIs both work
fedimintd reads ConfigGenSettings from env / defaults. Putting all these
variables in the fedimintd server opts is a little
isn't great. People will be confused if they change the values after
running config gen and the config doesn't update. But this seemed
cleaner for now because they're all segregated to a very discoverable
place in the code.