chore: protect legacy client high-level note-related operations
... from concurrent use

Just serialize with a mutex, bolted a bit ad-hoc
in there.

In seems that in the gateway we already handle incoming
payments on a separate task, without any protection,
and in the Fedi Alpha deployement we've noticed some strage
behavior where some outputs/notes seems to be missing etc.

This should also enable some concurrent usage of the `Client`,
serializing only what we know is not safe.