fix(peer): reconnection ping-pongs
The root of all evil is state. Especially mutable and shared. Seems to
me like peer communication is using a connection where both sides think
they are in control of the connection. So with higher latencies it is
possible to get into a sort of a reconnect ping-pong, when one sides
reconnects, starts re-sending all messages, and during that time the
other side reconnects, and starts sending messages themselves... just to
get new connection from the other side...

Fix #2800