fix: purge "block height", use "block count"
It has been established more than a half-century ago that
in computer science and engineering when operating on indexes, ranges
and counts it's best to start numbering from 0, and use half-open ranges:

https://www.cs.utexas.edu/users/EWD/transcriptions/EWD08xx/EWD831.html

Mathematicians sometimes disagree, but what do they know. And there's Lua,
which is a neat PL, yet most people using it complain precisely about the
1-based indexing (unless they are mathematicians).

Given that block height are 0-based and genesis block is has a height of 0:

https://mempool.space/block/000000000019d6689c085ae165831e934ff763ae46a2a6c172b3f1b60a8ce26f

it is most ergonomic to operate on "block count".

Bitcoin Core did the right thing and has a `getblockcount` call.

It's also worth noting that "block height" (as opposed to a
 "block count") is a misnomer. Which block? "Blockchain height" could
maybe work, so would "last/current/tip height", but
bare "block height" is just confusing.

There are situations in which "give me the *content* (or a hash)
of the last block" is a useful API to have, but even then a combination
of "get block count" and "get content/hash at height N" is a much
better API, unless that extra-call latency is really intolerable
(which is never).

I guess people "on the streets" (Twitters?), like to talk about
"we're at block height X" to meassure blockchain growth, but they
also count from 1, so yeah...