refactor!: Future proof arguments to ServerModuleInit::init
Backport of #3255

See #3356 for context.

It's a breaking change for the Rust code API for external
modules, but should be invisible for external users.

Original commit message:

Like #3254 but for server modules.
Re #3025

I'm not sure if this is not over-engineering, but it seems
likely that since client side needs it, server side
will need it too. We were already thinking about passing
`identity: PeerId` etc. through it.