refactor: `ClientWeak` and move logic to `ClientInner`
Rationale:

I'd like a `ClientWeak` to be a thing. Easy.

But then I'd also want a way for APIs to use `ClientWeak`, upgrade
it to `Client` (strong Arc), yet not allow the user of such APIs
to store them (to prevent introducing harmful ownership cycles).

For this to not require duplicating forwarding methods manually,
the inner `ClientInner` should just do as much as possible.
So in thie change I'm moving all methods to `ClientInner`,
making it more usable.

Also, now `Client` will just `Deref` to `ClientInner` to make the
whole thing mostly invisible from the user perspective. And
I think that's totally OK and even preferable. After all `Client`
is just a newtype-ed smart-pointer to `ClientInner`.

We can control the visibility of methods of `ClientInner` with
`pub`, `pub(crate)` etc. Importantly, this ends the confusing
and annoying split: some methods being implemented on `Client`,
some on `ClientInner`.

The only thing that bothers me is the naming now.