chore: get rid of `DatabaseTransactionRef`
A (hopefully) better approach than #3644.

Having a separate type for commitable vs non-commitable is a bit of
a pain that I've hit during refactoring, as there's no good way to
generalize over the two, and adding another layer of traits seems like
a nightmare.

This change unifies `DatabaseTransactionRef`
with `DatabseTransaction` using session types to express committability.

This makes generalizing overr `DatabaseTransaction` easier,
though it's not exactly perfect either. But I believe it's a bit better.