chore: add highest-level logs to sends_ecash_out_of_band
IMO we should be doing it in every test, so I'm going
to show it around.

When you are debugging a tricky problem by analyzing thousands
of logs lines of a failing test, having some reference points
telling you where the highest level points of test
execution happened is very, very useful.

Prefixing them with `###`, it makes jumping around and
finding them easier.