fix(memdb): waaay too slow, at least use rwlock
After a lot of debuging I discovered that some flakes are
due to all the slowness of `MemDatabase`.

On an even ligthly filled test database, coping all the data
start taking considerable amount of time (single digit milliseconds),
which combined with number of database transactions
being opened and commited, can lead to quite long delays in
processing in random places (as ordering of locking is somewhat
arbitrary).

It's possible that something is also running db transactions
in a tight loop somwhere.

For time being using an `RwLock` helps considerably, as
all db transactions at least need to start, and we do two
copied of the database content there, so being able to do it in parallel,
significantly improves the throughput.