feat(server): allow downloading encrypted guardian config
This allows restoring single, completely lost guardians
(DB and config gone). If fewer or equal to f guardians
are lost restoring can happen fairly automatic as shown
in the devimint test.

For more than f guardians we'd need to implement a
special procedure that lets the restoring ones trust
the remaining ones for the correctness of consensus
items from the ongoing session.