chore: add blank .semgrepignore & fix findings
If .semgrepignore is not present, semgrep uses a default .semgrepignore
file. The defaults excluded some tests which is not the desired outcome.

Add a blank .semgrepignore file to indicate that nothing additional to
.gitignore should be ignored by semgrep.

Also, add a second pattern to ban-direct-sleep-in-tests rule to find
the sleep function when the fully qualified function name is present.