chore: allocate client db prefix range for future use
Re #4444

I mentioned it before, but it was never followed up on.

Sometimes applications using client will want to store some data,
in client database. Potentially because it's already there, but
also due to atomic transaction requirements.

We should wrap it up in a nicer interface in the future, but for now,
just officially allocate it and document it in the code.

Not going to lie - while the range is somewhat arbitrary, and probably a smaller one would be enough, Fedi app has already started using this range, so it will make our life easier if we reserve it.

While I think one prefix would be enough, we (the Fedimint project) might come up with a need for other external use scenarios (you can't predict that stuff ahead of time), so might as well.

New users and app allocators are encuraged to just use the single
designated user data prefix.