fix: reject non-canonical encodings of BTreeMap/BTreeSet
This was found by the fuzzer.

This is OK to do because in practice our implementation always produced
canonical encodings only.

The bad part about the whole thing is that now `Ord` on `T` is consensus
significant. But this ship have sailed, as that's already the case in
the encoding part. If we were to start from scratch, we could consider
sorting they keys by their encoding instead.