chore: `meta` module
Meta module is meant to allow guardians to collectively setup and change over time a set of values for various auxiliary purposes, e.g.: establish customization of Federation like welcome messages, ToS and other legal documents, references to community information, chat servers, etc. This way the end user Fedimint client applications can discover these settings in a customize

It acts as a simple key-value store, where each guardian can submit a proposed value for a given key, and if `threshold` number of peers submit the same value, it becomes an official consensus.

This module is intentionally simple and limited in scope.

The keys are effectively a single byte, so there can only ever be 256 of them. Server side module does not interpret these values in any way and treats them as opaque binary blobs.

It is up to the community of Fedimint end user applications to establish the meaning and format of meta module values. The initial intention is to use a single key with the value of a json string. The structure of that JSON document is open-ended and can be estabilished and evolved over time, and applications can gracefully handle it in backward and forward compatible way.

This is all to move the social consensus changes outside of the technical consensus.

The extra 254 keys are an escape hatch - to avoid dealing with module consensus changes if a single value ever turns out to not be enough.