chore(devimint): speed up setup (pegins)
It looks like right now it takes longer peg-in the client & gateways
than to start all the servers and finish DKG. This is because
the peg-ins are done very naively.

We can in paralle deposit funds to gateways in clients, but
wait mine the blocks and await consensus single time. This cuts
the time significantly.