chore(client): speed up by merging "waiting for rejected" call
Use a small lru, just like we did before for merging module recovery
calls for session outcomes, to merge multiple calls for the same
acceptance/rejection query from multiple mint output state machines.

During mint bags creation, all notes are a different state
machine and there's a lot of them. However, they are all outputs
of the same transaction, so the result for all is going to be
exactly the same. Making dozens of same queries for exactly same
thing is a huge waste.

The mint output SM has two transition function there: one for
rejection, another for decryption shares (happy path). Ideally
we could merge both groups of calls, not just the one for rejection.
I'm thinking - if instead of asking for shares on each output,
we had some calls that returns shares for all outputs, we could
use the same technique as well.