chore(jit-values): return string-copy of the original error
Only the first caller to `Jit::get_try` can get the original error,
but we can actually return a string-copy of it to other callers, which
in most cases should be good enough.

The whole complication here allows retaining the original error
(including source chain, backtrace, etc.) which might be useful
when using in more concrete use-cases than just starting services
in Devimint.