chore: connect to peers in background (using Jit)
Since `request` is used in query strategies, it will often be canceled,
possibly cancelling the connection attempts. That's not great.

By using Jit, we can get connection to peers happen in the background,
and avoid the problem.

While at it - keep attempting to reconnect. The strategy will cancel
the request anyway.