refactor: GlobalApi should make non-admin calls to all peers
An-"admin-enabled" `GlobalApi` should behave for non-admin calls
just like non-"admin-enable" one. For this to happen self-peer-id
needs to be tracked separately instead of just narrowing `all_peers`
to one (self) peer_id.

In the future this should enable storing self-peer-id in the
Client's db, and using such client for both admin and non-admin
stuff all the same.

Re #4519