fix: `panic: Must deleted existing spendable note`
In one of the recent commits I've added an assertion to make sure a
spendable note was actually always removed when we thought it was,
as I was investigating a bug causing notes to be lost (#5195),
and it seemed like a good idea.

Too bad it causing flakes now: #5271

So it looks like our reliance on retrying db conflicts (`autocommit`),
leads to inability to assert basic expectations about database state,
which seems terrible.

In PR I'm adding a little lock to prevent write-write conflicts in this
area of the code, and I think it's worthwhile, just to be able to
ensure database integrity.

Moreover, I realized that `autocommit` is probably a bad idea altogether.

Application typically require transaction conflict handling, because
they are using **shared** databases. But with an embedded databases,
there is no sharing with external entities so there *always* should be,
or even *must be* a way to structure application and locking to avoid
any conflicts, preventing db isolation level issues, inability
to assert basic stuff, or just wasting time retrying things.