chore: do not require identical git hash during DKG
That is waay to constraining. Different peers can't even include
non-code fixes for their local deployement and other problems.

Since we're kind of good at the release and versioning and stability
things, change the requirement to a `code_version_str`, which
consists of the cargo release version of Fedimint code and optional
vendor suffix.