feat: add expert-user script for activating the meta module
Federations deployed prior to 0.3.0 do not have the meta module in their
config, meaning it doesn't get deactivated even though >=0.3.0 fedimintds
support it. We currently don't support adding modules after the fact, but in
case of the meta module we can kinda hack it in since it doesn't need any
config, which is what the added script does.