chore(ci): run a round of fuzzing for every fuzz target
Now that we self-hosted runners, that we don't pay for cycles used,
it makes sense to run some fuzzing all the time, just to avoid
accidental breakages and maybe even randomly discover something
over time.

It will be running with the same `runLowPrio` we use for other
CPU-heavy things, so it should not cause tests running in parallel
to be affected too much.