chore: future-proof api secrets in private federations
In the future we might want to manage secrets of a Federation
via UI, storing the Api Secrets configuration in the database.

This will make env var naming confusing, just it was in
case of a `FM_BITCOIN_RPC_URL`, etc. which we've split
into `FM_FORCE_` and `FM_DEFAULT_` version.

To the same for `FM_FORCE_API_SECRETS`, so the meaning
and behavior of it does not need to change in the future.

While at merge existing two arguments into one, to avoid
2 * 2 env vars/flags, and unify everything in a newtype.