refactor(consensus): some pass-by-ref tweaks
When working #5384 , which probably will get abandoned
I needed/noticed some ways to avoid a clone and passed
by reference stuff that didn't need to be moved.

The `&mut self` is not neccessary, but technically it
does denote that the consensus code has an exclusive reference
to `ConsensusEngine` which makese sense.