feat(server): diagnostic consensus checksum
Inspired by the discussion during last Deep Dive, about verifing
determinism of consensus between peers and shuting down when
incosistency was detected, it occured to me that we could get 90% there
by simply tracking a rolling checksum of what happened and displaying it
as diagnostics that guardians could sometimes simply compare.

So here is an initial implementation.

The checksum is maintained separately over txes and each of the modules,
so in case total checksum does not match it's possible to compare
each module and hopefully figure out which individual module is
to blame.