Remove usage of blockdata:: from bitcoin paths
In `rust-bitcoin` the `blockdata` module is a code organisation thing, it
should never have been public. One day those guys would like to remove
it, so as not to be a PITA for `fedimint` when they do lets remove
all usage of `blockdata::` now.

Internal change only, no externally visible changes.