refactor(client): resolve root_secret conflation
All over the code we call `root_secret` the level 0 secret derived from user/app
supplied secret hashed with the `FederationId`.

But in the `Client::open` and `Client::join` we call the secret before
the `FederationId` derivation the same name, which is very confusing.

In #5734 I decided to call the secret used to derive the `root_secret`
a "pre-root secret", and I think that kind of works, so might as well
rename it in these key places.