fix: esplora bitcoind backend returning reverse hashes
It seems to me that what happened is that during rust-bitcoin
upgrade the internal representation of `BlockHash` was reversed
and now naive `hex` crate encoding gives us wrong hash, that
is interpreted as regtest network. Using `.to_string` on
`BlockHash` does the right thing independent of underlying
order of bytes.