chore(client): increase size of the session fetching cache LRUs
I did some test recoveries on the real Federations and I observed,
that due to small cache size wallet & mint recovery fetching diverged
and was fetching sessions twice. Not the end of the world, but
the whole point of this cache was to avoid doube-downloading, so
a waste.

Increasing the size of the LRU, gives enough buffer for everything
to reliably stay in sync.

(cherry picked from commit 7effa6fc35f492dca21c66b9c59345be4a09097d)