fix: find_by_prefix should guarantee order
It's already the case, yet in the interface documentation we don't commit to it. I'm quite sure there are already parts of the code
that depend on it anyway.

In a system like ours any form of non-determism is unwelcomed.
I rather have the db impls have to do the sorting, than all users
need to sort. Especially that for all kv stores, sorting should
come naturally.