chore(client): check peg-in addresses with no deposits more often
Currently, if the user creates a deposit address, and then decides to
make a peg-in after a week, they might wait 3.5 day for the client to
notice.

Ultimately end-user apps needs to expose ability to check for deposits
on demand (click a button or something), but irrespective of that,
checking a 7 day old address arond once a day seems reasonable.