chore(wallet-server): don't spam citems
In theory wallet-server proposing citems all the time shouldn't matter,
as redudant citems will get discarded

In practice it is spamming logs and is simply unnecessary.
It also makes the peer start after restart slower, as it has to re-process (re-reject as redundant) all the items from the currently opened
session.

There are 3 things that cause us to want to propose a quick citem:

* block count (periodic)
* fee rate (periodic)
* tx signature ci

This change adds an additional watch channel to make the
wallet-server propose things when any of these 3 sent an update. This
will lower rate of citem proposals from ~2 ever second per peer to ~1
every 15-30 seconds.