fix: change/improve our network settings to accommodate iroh
Fundamentally Iroh typically needs to be bound to 0.0.0.0, while our
http apis to 127.0.0.1. We need two different settings for binding it,
as we ran both APIs at the same time (at least sometimes).

We can set the defaults to typical production settings:

* p2p (iroh/tls): 0.0.0.0
* iroh api: 0.0.0.0
* ws api: 127.0.0.1

Add some documentation with mention of typical settings, so `--help`
is more informative.

Fix our docker examples, as at least Iroh one was setting iroh
to bind to 127.0.0.1, which probably only worked because of fallback
TCP connections to public relays.