chore: try i18n using rust-i18n
I did a bit of searching for i18n and `rust-i18n` LGTM.

https://github.com/longbridge/rust-i18n/

It's compile time based, so less moving parts.

It comes down to wrapping a text in `t!(...)` macro, and then
maintaining translations in a yaml file.

It has a simple CLI tool that can scan project for missing translations.
ATM it seems to require `cd`-ing to a given crate first.


You can try this change with:

```
env FM_UI_LOCALE=pl just devimint-env-pre-dkg
```

In the future, we could use it for translating user errors as well.